rename structure
